+++ EasyVCR.Tests/Sample.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace EasyVCR.Tests
@@ -5,21 +6,28 @@ namespace EasyVCR.Tests
     public class Sample
     {
         /// <summary>
-        ///     Create a VCR, create a cassette, insert cassette into VCR, get the HttpClient from the VCR and make a request.
+        ///     Set advanced options for a cassette, applied to all requests made using the cassette.
+        ///     Set specific order options when creating a cassette.
         /// </summary>
-        public async Task SimpleVCRExample()
+        public async Task AdvancedCassetteExample()
         {
-            var vcr = new VCR();
-            var cassette = new Cassette("path/to/cassettes", "my_cassette");
-            vcr.Insert(cassette);
-            vcr.Record();
+            var advancedSettings = new AdvancedSettings
+            {
+                MatchRules = MatchRules.DefaultStrict, // use the built-in strict match rules
+                Censors = Censors.DefaultSensitive, // use the built-in sensitive censors
+                SimulateDelay = true, // simulate the exact delay of the original request during playback
+                // InteractionConverter = new MyInteractionConverter(), // use a custom interaction converter by implementing IInteractionConverter
+            };
+            var order = new CassetteOrder.Alphabetical(); // elements of each request in a cassette will be ordered alphabetically
 
-            var httpClient = vcr.Client;
+            var cassette = new Cassette("path/to/cassettes", "my_cassette", order);
+
+            var httpClient = HttpClients.NewHttpClient(cassette, Mode.Auto, advancedSettings);
 
             // Use the httpClient as you would normally.
             var response = await httpClient.GetAsync("https://google.com");
 
-            vcr.Eject();
+            cassette.Lock();
         }
 
         /// <summary>
@@ -31,7 +39,7 @@ namespace EasyVCR.Tests
             var advancedSettings = new AdvancedSettings
             {
                 MatchRules = new MatchRules().ByBody().ByHeader("X-My-Header"), // Match recorded requests by body and a specific header
-                Censors = new Censors("redacted").HideHeader("Header-To-Hide").HideQueryParameter("api_key"), // Redact a specific header and query parameter 
+                Censors = new Censors("redacted").HideHeaders(new List<string> { "Header-To-Hide" }).HideQueryParameters(new List<string> { "api_key" }), // Redact a specific header and query parameter 
                 ManualDelay = 1000, // Simulate a delay of 1 second
             };
             var order = new CassetteOrder.None(); // elements of each request in a cassette will not be ordered any particular way
@@ -64,28 +72,21 @@ namespace EasyVCR.Tests
         }
 
         /// <summary>
-        ///     Set advanced options for a cassette, applied to all requests made using the cassette.
-        ///     Set specific order options when creating a cassette.
+        ///     Create a VCR, create a cassette, insert cassette into VCR, get the HttpClient from the VCR and make a request.
         /// </summary>
-        public async Task AdvancedCassetteExample()
+        public async Task SimpleVCRExample()
         {
-            var advancedSettings = new AdvancedSettings
-            {
-                MatchRules = MatchRules.DefaultStrict, // use the built-in strict match rules
-                Censors = Censors.DefaultSensitive, // use the built-in sensitive censors
-                SimulateDelay = true, // simulate the exact delay of the original request during playback
-                // InteractionConverter = new MyInteractionConverter(), // use a custom interaction converter by implementing IInteractionConverter
-            };
-            var order = new CassetteOrder.Alphabetical(); // elements of each request in a cassette will be ordered alphabetically
+            var vcr = new VCR();
+            var cassette = new Cassette("path/to/cassettes", "my_cassette");
+            vcr.Insert(cassette);
+            vcr.Record();
 
-            var cassette = new Cassette("path/to/cassettes", "my_cassette", order);
-
-            var httpClient = HttpClients.NewHttpClient(cassette, Mode.Auto, advancedSettings);
+            var httpClient = vcr.Client;
 
             // Use the httpClient as you would normally.
             var response = await httpClient.GetAsync("https://google.com");
 
-            cassette.Lock();
+            vcr.Eject();
         }
     }
 }