+++ EasyVCR/Properties/VersionInfo.cs
@@ -1,6 +1,6 @@
 ﻿using System.Reflection;
 
 // Version information for an assembly must follow semantic versioning
-[assembly: AssemblyVersion("0.2.0")]
-[assembly: AssemblyFileVersion("0.2.0")]
-[assembly: AssemblyInformationalVersion("0.2.0")]
+[assembly: AssemblyVersion("0.3.0")]
+[assembly: AssemblyFileVersion("0.3.0")]
+[assembly: AssemblyInformationalVersion("0.3.0")]