+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # CHANGELOG
 
-## NEXT RELEASE
+## v0.3.0 (2022-05-24)
 
-- Body censors apply to nested JSON body keys, not just top-level keys
-  - Lists replaced with empty lists, dictionaries replaced with empty dictionaries, primitives replaced with censor text
+- Performance improvements to censoring
+  - Censoring process is skipped if there are no censors to apply
+  - Body censors apply to nested JSON body keys, not just top-level keys
+    - Lists replaced with empty lists, dictionaries replaced with empty dictionaries, primitives replaced with censor text
 - Censor declaration functions expect lists of multiple keys
 
 ## v0.2.0 (2022-05-18)